+++ app/src/main/java/com/monero/models/Credit.kt
@@ -14,4 +14,4 @@ data class Credit(@ColumnInfo(name = "credit_id") @PrimaryKey val id:Long,
                   val expense_id: String,
                   val user_id: String,
                   val userName:String,
-                  val amount:Double)
+                  val amount: Int)
